INS-614: Move all data manipulation routines to logicrunner/accessors.go
